fix cursor placement for indented php tag
expanding `body>php` would leave the cursor in the 2nd column inside the
php tag instead of at the end of the line. using `_` for the cursor position and periods for trailing spaces, the expansion looked like:

```
<body>
    <?php
 _......
    ?>
</body>
```

ensure that `s:SparkupNext` correctly returns the matching subpattern for an
empty line.