Fix reading of input and handling of Ctrl-C
The catch for "Reading failed." is not neccessary in my opinion.
If sparkup is used like
  sparkup < file
and file cannot be opened, the shell will give an error.
For sparkup itself, it should never be a problem to read from stdin.

Catching KeyboardInterrupt here was an error: the variable "lines" is used
later on and it will be undefined when Ctrl-C is pressed.

  ^CParse error. Check your input.
  <class 'UnboundLocalError'>
  local variable 'lines' referenced before assignment

I would just wrap the try-except for KeyboardInterrupt around the main
code.